insurance_bot_agents/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insurance-bot-agents"
version = "1.0.0"
description = "Insurance Bot Agents - A collection of AI agents for insurance domain tasks"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["insurance", "ai-agents", "langchain", "langgraph", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # AI/ML Frameworks
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.9",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.68",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "langgraph-swarm>=0.0.12",
    "langsmith>=0.4.5",

    # OpenAI & Anthropic
    "openai>=1.95.1",
    "anthropic>=0.49.0",

    # FastAPI & Web
    "fastapi>=0.115.11",

    # Database
    "pymongo>=4.11.3",
    "sqlalchemy>=2.0.38",
    "psycopg2-binary>=2.9.10",

    # Data Processing
    "pandas>=2.2.3",
    "numpy>=2.2.3",
    "pandasql>=0.7.3",

    # Visualization
    "matplotlib>=3.10.1",
    "seaborn>=0.13.2",

    # MCP Tools
    "mcp>=1.10.1",
    "fastmcp>=2.12.4",

    # Utilities
    "python-dotenv>=1.1.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.8.1",
    "httpx>=0.28.1",
    "tenacity>=9.1.2",

    # Vector Search
    "qdrant-client>=1.15.1",
    "langchain-qdrant>=0.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/insurance-bot-agents"
Repository = "https://github.com/yourusername/insurance-bot-agents"

[tool.setuptools]
packages = ["insurance_bot_agents"]
package-dir = {"insurance_bot_agents" = "."}

[tool.setuptools.package-data]
insurance_bot_agents = ["*.txt", "*.md"]
